--- src/Components/Feed/feed.css
@@ -1,6 +1,0 @@
-.divAdd{
-    position: relative;
-    bottom: 10%;
-    right:0;
-    padding: 24000;
-}
