+++ src/Components/Search/paper.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
+import { Grid} from '@material-ui/core' ;
 
 const styles = theme => ({
   root: {
@@ -17,7 +18,8 @@ function PaperSheet(props) {
 
   return (
     <div>
-      <Paper className={classes.root} elevation={1}>
+      <Paper className={classes.root}>
+      <Grid container direction="column" justify="center" alignItems="center">
         <Typography component="p">
           Tema: Calculo 1
         </Typography>
@@ -27,6 +29,7 @@ function PaperSheet(props) {
         <Typography component="p">
           Data: 25/04/2020
         </Typography>
+        </Grid>
       </Paper>
     </div>
   );
